+++ Runtime/!Core/CompilerServices.cs
@@ -1,0 +1,5 @@
+// This is some magic code to enable the init and record keywords in Unity.
+// This is required because Unity uses an older version of C# that does not support these keywords.
+namespace System.Runtime.CompilerServices {
+  class IsExternalInit { }
+}