+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Any notable changes to this project will be documented in this file.
 Initial release of Shadowpaw - Unity Toolkit
 
 - `License`: Changed to the [UNLICENSE](https://unlicense.org) (Public Domain)
+- `Core`: Added some magic code to enable the init and record keywords in Unity
 - `Utils`: Added extensions for LayerMask, List, Reflection, String, Task, and Type
 
 Be good to each other 💕